Easier getting-started experience
### Overview
This readme provides the reader with an easier getting-started experience. In other words, it is now _faster_ to _plug-and-play_ the credentialdigger and get directly to the scanning point without the need to dig deeper into the docs nor to set up a Postgres instance.
It also helps those who want a to get a working instance simply by doing CTRL-C & CTRL-V without further *digging* :-)

On a side not, the `rules` wiki page has been updated.